+++ routes/admin/debate.js
@@ -1,25 +1,10 @@
 const express = require('express');
 
 const router = express.Router();
-const dynamoDb = require('../models/dynamoDb');
-const { getS3oUsername } = require('../helpers/cookies');
+const dynamoDb = require('../../models/dynamoDb');
+const { getS3oUsername } = require('../../helpers/cookies');
 
-router.get('/', async (req, res) => {
-  const username = getS3oUsername(req.cookies);
-
-  try {
-    const debateList = await dynamoDb.getAllDebateLists();
-    res.render('admin/index', {
-      username,
-      debateList,
-      page: 'dashboard',
-    });
-  } catch (err) {
-    res.status(404).send("Sorry can't find that!");
-  }
-});
-
-router.get('/create_debate', async (req, res) => {
+router.get('/create', async (req, res) => {
   try {
     const username = getS3oUsername(req.cookies);
     const debateTypes = await dynamoDb.getAllDebateTypes();
@@ -34,7 +19,7 @@ router.get('/create_debate', async (req, res) => {
   }
 });
 
-router.post('/create_debate', async (req, res) => {
+router.post('/create', async (req, res) => {
   try {
     const {
       debateType,
@@ -67,7 +52,7 @@ router.post('/create_debate', async (req, res) => {
   }
 });
 
-router.get('/edit_debate/:debateUuid', async (req, res) => {
+router.get('/edit/:debateUuid', async (req, res) => {
   const username = getS3oUsername(req.cookies);
 
   try {
@@ -117,7 +102,7 @@ router.get('/edit_debate/:debateUuid', async (req, res) => {
   }
 });
 
-router.post('/edit_debate/:uuid', async (req, res) => {
+router.post('/edit/:uuid', async (req, res) => {
   const username = getS3oUsername(req.cookies);
   try {
     const { uuid } = req.params;
@@ -153,97 +138,18 @@ router.post('/edit_debate/:uuid', async (req, res) => {
   }
 });
 
-router.get('/moderation', async (req, res) => {
-  const username = getS3oUsername(req.cookies);
-
-  try {
-    const reports = await dynamoDb.getAllReports();
-
-    res.render('admin/moderation', {
-      username,
-      reports,
-      page: 'moderation',
-    });
-  } catch (err) {
-    res.status(404).send("Sorry can't find that!");
-  }
-});
-
-router.get('/create_debate_type', (req, res) => {
-  const username = getS3oUsername(req.cookies);
-
-  res.render('admin/createDebateType', {
-    username,
-    page: 'create-type',
-  });
-});
-
-router.post('/create_debate_type', async (req, res) => {
-  try {
-    const { specialUsers, name, description, displayName } = req.body;
-    const defaultSpecialUser = [
-      {
-        name: 'allowedUsers',
-        description: 'Add users to a allow list.',
-        displayName: 'Allowed Users',
-      },
-      {
-        name: 'blockedUsers',
-        description: 'Add users to a block list.',
-        displayName: 'Blocked Users',
-      },
-    ];
-    const result = await dynamoDb.createDebateType({
-      specialUsers: [...defaultSpecialUser, ...specialUsers],
-      name,
-      description,
-      displayName,
-    });
-    const username = getS3oUsername(req.cookies);
-    res.render('admin/createDebateType', {
-      username,
-      success: true,
-      page: 'create-type',
-    });
-  } catch (err) {
-    console.error(err);
-  }
-});
-
-router.get('/edit_debate_type/:debateName', async (req, res) => {
-  try {
-    const username = getS3oUsername(req.cookies);
-    const { debateName } = req.params;
-    const debateType = await dynamoDb.getDebateType(debateName);
-    if (debateType.Items.length === 0) {
-      throw new Error('Cant find debate type');
+function formatSpecialUsers(specialUsers) {
+  let spcialUsersFormatted = [];
+  Object.keys(specialUsers).forEach(userType => {
+    if (typeof specialUsers[userType] === 'string') {
+      specialUsers[userType] = [specialUsers[userType]];
     }
-
-    const {
-      description,
-      name,
-      specialUsers,
-      displayName,
-    } = debateType.Items[0];
-
-    res.render('admin/editDebateType', {
-      username,
-      description,
-      name,
-      displayName,
-      specialUsers: specialUsers.map((specialUser, index) => ({
-        ...specialUser,
-        index,
-      })),
-      page: 'edit-type',
-    });
-  } catch (err) {
-    console.error(err);
-    res.render('admin/editDebateType', {
-      error: 'Something went wrong retrieving your debate type',
-      page: 'edit-type',
-    });
-  }
-});
+    spcialUsersFormatted = [
+      ...spcialUsersFormatted,
+      { userType, users: specialUsers[userType] },
+    ];
+  });
+  return spcialUsersFormatted;
+}
 
 module.exports = router;
